Remove obsolete benchmark
