+++ exonum/benches/criterion/storage.rs
@@ -99,6 +99,49 @@ fn merkle_patricia_table_merge_insert<T: Database>(b: &mut Bencher, db: &T) {
     });
 }
 
+fn proof_map_index_build_proofs<T: Database>(b: &mut Bencher, db: &T) {
+    let data = generate_random_kv(512);
+    let mut storage = db.fork();
+    let mut table = ProofMapIndex::new(NAME, &mut storage);
+
+    for item in &data {
+        table.put(&item.0, item.1.clone());
+    }
+    let table_merkle_root = table.merkle_root();
+    let mut proofs = Vec::with_capacity(data.len());
+
+    b.iter(|| {
+        proofs.clear();
+        proofs.extend(data.iter().map(|item| table.get_proof(item.0)));
+    });
+
+    for (i, proof) in proofs.into_iter().enumerate() {
+        let checked_proof = proof.check().unwrap();
+        assert_eq!(*checked_proof.entries().next().unwrap().1, data[i].1);
+        assert_eq!(checked_proof.merkle_root(), table_merkle_root);
+    }
+}
+
+fn proof_map_index_verify_proofs<T: Database>(b: &mut Bencher, db: &T) {
+    let data = generate_random_kv(512);
+    let mut storage = db.fork();
+    let mut table = ProofMapIndex::new(NAME, &mut storage);
+
+    for item in &data {
+        table.put(&item.0, item.1.clone());
+    }
+    let table_merkle_root = table.merkle_root();
+    let proofs: Vec<_> = data.iter().map(|item| table.get_proof(item.0)).collect();
+
+    b.iter(|| {
+        for (i, proof) in proofs.iter().enumerate() {
+            let checked_proof = proof.clone().check().unwrap();
+            assert_eq!(*checked_proof.entries().next().unwrap().1, data[i].1);
+            assert_eq!(checked_proof.merkle_root(), table_merkle_root);
+        }
+    });
+}
+
 fn create_rocksdb(tempdir: &TempDir) -> RocksDB {
     let options = DbOptions::default();
     RocksDB::open(tempdir.path(), &options).unwrap()
@@ -122,16 +165,36 @@ fn bench_merkle_patricia_table_insertion_fork_rocksdb(b: &mut Bencher) {
     merkle_patricia_table_fork_insert(b, &db);
 }
 
+fn bench_proof_map_index_build_proofs_rocksdb(b: &mut Bencher) {
+    let tempdir = TempDir::new("exonum").unwrap();
+    let db = create_rocksdb(&tempdir);
+    proof_map_index_build_proofs(b, &db);
+}
+
+fn bench_proof_map_index_validate_proofs_rocksdb(b: &mut Bencher) {
+    let tempdir = TempDir::new("exonum").unwrap();
+    let db = create_rocksdb(&tempdir);
+    proof_map_index_verify_proofs(b, &db);
+}
+
 pub fn bench_storage(c: &mut Criterion) {
     ::exonum::crypto::init();
 
     c.bench_function("append merkle table", bench_merkle_table_append_rocksdb);
     c.bench_function(
-        "insert without merge merkle table",
+        "insert without merge merkle patricia table",
         bench_merkle_patricia_table_insertion_fork_rocksdb,
     );
     c.bench_function(
-        "insert with merge merkle table",
+        "insert with merge merkle patricia table",
         bench_merkle_patricia_table_insertion_merge_rocksdb,
+    );
+    c.bench_function(
+        "build proofs merkle patricia table",
+        bench_proof_map_index_build_proofs_rocksdb,
+    );
+    c.bench_function(
+        "validate proofs merkle patricia table",
+        bench_proof_map_index_validate_proofs_rocksdb,
     );
 }
